Improve HexaLock AutoLock detection
Vastly improve HexaLock AutoLock detection, and add notes.
